[Feature] Adding NoItemsView

+++ TodoList/Views/ListView.swift
@@ -7,28 +7,32 @@
 
 import SwiftUI
 
-
- 
-
 struct ListView: View {
     
     @EnvironmentObject var listViewModel: ListViewModel
     
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear){
-                            listViewModel.updateItem(item: item)
-                        }
+        ZStack {
+            if listViewModel.items.isEmpty {
+                NoItemsView()
+                    .transition( AnyTransition.opacity.animation(.easeIn))
+            } else {
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear){
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
+                .listStyle(PlainListStyle())
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
         }
-        .listStyle(PlainListStyle())
         .navigationTitle(Text("Todo List 📝"))
         .navigationBarItems(
             leading: EditButton(),
@@ -36,7 +40,6 @@ struct ListView: View {
                 NavigationLink("Add", destination: AddView())
            )
     }
-
 }
 
 #Preview {
